Move existing pipeline to companies and add pipeline tab
Make the folder structure consistent with the routes
The pipleline on the dashboard is at /my-pipleline
The adding of a company to a pipleline is from within the companies sub-app

Add a new pipeline component to with placeholder text
