+++ src/apps/companies/apps/pipeline/client/reducer.js
@@ -1,7 +1,7 @@
 import {
   PIPELINE__CHECKED_IF_ON_PIPELINE,
   PIPELINE__ADD_COMPANY_SUCCESS,
-} from '../../../client/actions'
+} from '../../../../../client/actions'
 
 export default (state = {}, { type, result }) => {
   switch (type) {
