+++ src/apps/companies/apps/pipeline/client/state.js
