+++ src/apps/companies/apps/pipeline/index.js
@@ -1,0 +1,8 @@
+const urls = require('../../../../lib/urls')
+
+const router = require('./router')
+
+module.exports = {
+  mountpath: urls.companies.pipeline.mountPoint,
+  router,
+}
