--- src/apps/pipeline/index.js
@@ -1,8 +1,0 @@
-const urls = require('../../lib/urls')
-
-const router = require('./router')
-
-module.exports = {
-  mountpath: urls.pipeline.index.mountPoint,
-  router,
-}
