+++ src/client/components/Pipeline/index.jsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function Pipeline() {
+  return <p>My pipeline content</p>
+}
